CAMMI - Team/Assignment Task 2/requirements.txt
python-docx
torch
transformers
accelerate
sentencepiece